refactor(web): classify loose components into atomic layers
Move the 7 unclassified components out of src/components root into their
proper Atomic Design layers, so the structure is consistent and discoverable:
- organisms/: emergency-map, emergency-map-wrapper, leaflet-map, location-picker
  (kept together so the leaflet dynamic ssr:false imports keep resolving)
- molecules/: needs-filter, org-selector, social-login-buttons
Imports updated across pages + forms. lint + build pass.
