+++ apps/web/src/app/e/[slug]/donar/page.tsx
@@ -3,8 +3,8 @@ import { notFound, redirect } from 'next/navigation';
 import { getEmergencyBySlug } from '@/lib/emergencies';
 import { getToken } from '@/lib/auth';
 import { api } from '@/lib/api';
-import { OrgSelector } from '@/components/org-selector';
-import { LocationPicker } from '@/components/location-picker';
+import { OrgSelector } from '@/components/molecules/org-selector';
+import { LocationPicker } from '@/components/organisms/location-picker';
 import { submitOffer } from './actions';
 import { DonarForm } from './donar-form';
 import { PageHeaderBand } from '@/components/molecules/page-header-band';
