+++ apps/web/src/app/e/[slug]/page.tsx
@@ -5,8 +5,8 @@ import { getEmergencyBySlug } from '@/lib/emergencies';
 import { getToken } from '@/lib/auth';
 import { OfficialHeaderBand } from '@/components/organisms/official-header-band';
 import { ResourceList } from '@/components/organisms/resource-list';
-import { EmergencyMapWrapper } from '@/components/emergency-map-wrapper';
-import { NeedsFilter } from '@/components/needs-filter';
+import { EmergencyMapWrapper } from '@/components/organisms/emergency-map-wrapper';
+import { NeedsFilter } from '@/components/molecules/needs-filter';
 import { EmptyState } from '@/components/molecules/empty-state';
 import { MetricCard } from '@/components/molecules/metric-card';
 import { StatusBanner } from '@/components/molecules/status-banner';
@@ -14,7 +14,7 @@ import { AnnouncementCard } from '@/components/molecules/announcement-card';
 import { HelpActionRow } from '@/components/molecules/help-action-row';
 import { NeedCard } from '@/components/molecules/need-card';
 import { EmergencyQuickLinks } from '@/components/molecules/emergency-quick-links';
-import type { MapPoint } from '@/components/emergency-map';
+import type { MapPoint } from '@/components/organisms/emergency-map';
 import { getT } from '@/i18n/server';
 
 export const dynamic = 'force-dynamic';
