+++ apps/web/src/app/e/[slug]/peticion/page.tsx
@@ -2,8 +2,8 @@ import type { Metadata } from 'next';
 import { notFound, redirect } from 'next/navigation';
 import { getEmergencyBySlug } from '@/lib/emergencies';
 import { getToken } from '@/lib/auth';
-import { OrgSelector } from '@/components/org-selector';
-import { LocationPicker } from '@/components/location-picker';
+import { OrgSelector } from '@/components/molecules/org-selector';
+import { LocationPicker } from '@/components/organisms/location-picker';
 import { submitPeticion } from './actions';
 import { PeticionForm } from './peticion-form';
 import { ItemsField } from './items-field';
